+++ src/RoadRegistry.Syndication.Schema/MigrationTables.cs
@@ -1,7 +1,0 @@
-namespace RoadRegistry.Syndication.Schema
-{
-    public static class MigrationTables
-    {
-        public const string Syndication = "__EFMigrationsHistorySyndication";
-    }
-}