+++ src/RoadRegistry.Wms.Schema/MigrationTables.cs
@@ -1,7 +1,0 @@
-namespace RoadRegistry.Wms.Schema
-{
-    public static class MigrationTables
-    {
-        public const string Wms = "__EFMigrationsHistoryWms";
-    }
-}